Numerous fixes to config_from_env
* Move to separate module
* `env` argument may be omitted, defaults to `os.environ`
* Add workaround for `str.removeprefix`, which was added in Python 3.9
* Move over unit tests, enable test suite in CI pipeline
