+++ .github/workflows/tox.yml
@@ -8,6 +8,7 @@ jobs:
   test:
     uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@v1
     with:
+      coverage: 'codecov'
       envs: |
         - linux: py37
         - linux: py38
