+++ .gitignore
@@ -1,5 +1,6 @@
 *.egg-info
 __pycache__
+.coverage
 .tox
 .vscode
 build
