+++ gcn_kafka/__init__.py
@@ -1,6 +1,7 @@
 # SPDX-License-Identifier: CC0-1.0
 
 from .core import Consumer, Producer
+from .env import config_from_env
 from ._version import version as __version__
 
-__all__ = ("Consumer", "Producer")
+__all__ = ("config_from_env", "Consumer", "Producer")
