+++ setup.cfg
@@ -23,3 +23,7 @@ install_requires =
     typing-extensions; python_version<='3.7'
 packages = find:
 python_version = >=3.7
+
+[options.extras_require]
+test =
+    pytest
