+++ tox.ini
@@ -3,4 +3,9 @@ envlist = py{37,38,39}
 isolated_build = True
 
 [testenv]
-commands = python -c 'import gcn_kafka'
+deps =
+    pytest-cov
+extras =
+    test
+commands =
+    pytest gcn_kafka --cov
